requirements.txt
fastapi
uvicorn
python-jose[cryptography]
requests
psycopg2-binary
google-generativeai
streamlit
ollama
python-dotenv
